Gmail_Agent/my_agent/requirements.txt
langchain-openai>=0.0.2
langchain-core>=0.1.0
langchain-community>=0.0.2
langgraph>=0.0.15
openai>=1.3.7
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.0.0
qdrant-client>=1.4.0
python-dotenv>=1.0.0
